requirements.txt
openai>=1.42.0
requests>=2.32.3
python-dotenv>=1.0.1
Pillow>=10.4.0
tqdm>=4.66.5
PyJWT>=2.8.0
requests-toolbelt>=1.0.0
google-genai>=0.3.0
flask>=3.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.11.0
boto3>=1.28.0
